Reduced add and remove route.ts files into one route.ts file in the game/[gameID] folder, and made it so the GameInLibraryInfo.tsx file uses this custom route for checking the library status.
